+++ lib/academic_year.rb
@@ -1,0 +1,42 @@
+# Used to model an academic year, which is normally dispayed in the format
+# "YYYY/YYYY". Supports comparison and basic arithmetic operations.
+#
+# Can be initialised both with a single start year, or the academic year as a
+# string. For example:
+#
+#   AcademicYear.new(2014).to_s        #=> "2014/2015"
+#   AcademicYear.new("2014/2015").to_s #=> "2014/2015"
+#
+# It also supports being displayed in a user-friendly format:
+#
+#   AcademicYear.new("2014/2015").to_s(:long) #=> "2014 to 2015"
+class AcademicYear
+  include Comparable
+
+  attr_reader :start_year, :end_year
+
+  def initialize(start_year)
+    @start_year = start_year.to_s.split("/").first.to_i
+    @end_year = @start_year + 1
+  end
+
+  def to_s(format = :default)
+    if format == :long
+      "#{start_year} to #{end_year}"
+    else
+      "#{start_year}/#{end_year}"
+    end
+  end
+
+  def <=>(other)
+    start_year <=> other.start_year
+  end
+
+  def -(other)
+    AcademicYear.new(start_year - other)
+  end
+
+  def +(other)
+    AcademicYear.new(start_year + other)
+  end
+end