+++ spec/mailers/claim_mailer_spec.rb
@@ -72,6 +72,18 @@ RSpec.describe ClaimMailer, type: :mailer do
         it "mentions that claim has been rejected in the subject and body" do
           expect(mail.subject).to include("rejected")
           expect(mail.body.encoded).to include("not been able to approve")
+
+          ineligible_year = (policy.first_eligible_qts_award_year - 1).to_s(:long)
+          expect(mail.body.encoded)
+            .to include("completed your initial teacher training in or before the academic year #{ineligible_year}")
+        end
+
+        it "changes the ITT reason based on the policy's configured current_academic_year" do
+          PolicyConfiguration.for(policy).update!(current_academic_year: "2025/2026")
+
+          ineligible_year = (policy.first_eligible_qts_award_year - 1).to_s(:long)
+          expect(mail.body.encoded)
+            .to include("completed your initial teacher training in or before the academic year #{ineligible_year}")
         end
       end