+++ spec/models/maths_and_physics/admin_checks_presenter_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe MathsAndPhysics::AdminChecksPresenter, type: :model do
       current_school: school,
       initial_teacher_training_subject: :maths,
       initial_teacher_training_subject_specialism: nil,
-      qts_award_year: "on_or_after_september_2014")
+      qts_award_year: "on_or_after_cut_off_date")
   end
   subject(:presenter) { described_class.new(eligibility) }