+++ spec/models/student_loans/admin_checks_presenter_spec.rb
@@ -5,7 +5,7 @@ RSpec.describe StudentLoans::AdminChecksPresenter, type: :model do
   let(:eligibility) do
     build(
       :student_loans_eligibility,
-      qts_award_year: "on_or_after_september_2013",
+      qts_award_year: "on_or_after_cut_off_date",
       claim_school: school,
       current_school: school,
     )